==> Step 3 - Build Topic Models.ipynb <==
import newspaper
import feedparser
import numpy as np
import pandas as pd
import requests
import datetime 
from tqdm import tqdm
import nltk
from nltk.corpus import stopwords
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.decomposition import NMF, LatentDirichletAllocation
import joblib
import spacy

## Topic Modeling


## Inputs and outputs
There are just two cells for doing topic modeling.  The first one  
specifies the different inputs.  Adjust these to what we need  
the topic modeling to do, ie number of topics, top n words per  
topic that we wish to see, count vs tfidf, and Non-negative  
Matrix Factorization vs Latent Dirichlet Algorithm.  
  
Once done, the below code will create with two dataframes:  
Main output is:  
 - words_in_topics_df - top_n_words per topic  
 - topic_for_doc_df - topic to which a document is identified  
  
Additional outputs of interest:  
  
 - vocab = This is the dict from which you can pull the words, eg vocab['ocean']  
 - terms = Just the list equivalent of vocab, indexed in the same order  
 - doc_term_matrix = Document term matrix   
Now doc_term_matrix is factorized as = W x H.  You can get W and H:    
 - W = This matrix has docs as rows and num_topicss as columns  
 - H = This matrix has num_topics as rows and vocab as columns  
  
***  



final = joblib.load('final_df3.pkl')

# Read stopwords from file
custom_stop_words = []
file = open(file = "stopwords.txt", mode = 'r')
custom_stop_words = file.read().split('\n')


# # How to write to a file
# # Uncomment everything in this cell to write to a file
# file = open('stopwords.txt','w')
# for element in custom_stop_words:
#     file.write(element+'\n')

# file.close()

# Input incoming text as a list called raw_documents

raw_documents= list(final['text'])
# custom_stop_words = joblib.load('stopwords.pkl')
# custom_stop_words = stopwords.words('english') - use this from nltk if custom file not available
num_topics = 10
top_n_words = 20
vectorizer_to_use = 1 # Use 1 for CountVectorizer, and 2 for TFIDF_Vectorizer
NMF_or_LDA = 'nmf' # Use 'nmf' for NMF or 'lda' for LDA
ngram = 2 # 2 for bigrams, 3 for trigrams etc

# Once done, the below code will create with two dataframes:
# Main output is:
#     words_in_topics_df - top_n_words per topic
#     topic_for_doc_df - topic to which a document is identified

# Additional outputs of interest
# vocab = This is the dict from which you can pull the words, eg vocab['ocean']
# terms = Just the list equivalent of vocab, indexed in the same order
# term_frequency_table = dataframe with the frequency of terms
# doc_term_matrix = Document term matrix 
# doc_term_matrix = W x H
# W = This matrix has docs as rows and num_topicss as columns
# H = This matrix has num_topics as rows and vocab as columns

    

# use count based vectorizer
if vectorizer_to_use ==1:
    vectorizer = CountVectorizer(stop_words = custom_stop_words, min_df = 2, analyzer='word', ngram_range=(ngram, ngram))
else:
    # or use TF-IDF based vectorizer
    vectorizer = TfidfVectorizer(max_df=0.95, min_df=2, stop_words=custom_stop_words, analyzer='word', ngram_range=(ngram, ngram))

doc_term_matrix = vectorizer.fit_transform(raw_documents)
print( "Created %d X %d document-term matrix in variable doc_term_matrix\n" % (doc_term_matrix.shape[0], doc_term_matrix.shape[1]) )




vocab = vectorizer.vocabulary_ #This is the dict from which you can pull the words, eg vocab['ocean']
terms = vectorizer.get_feature_names() #Just the list equivalent of vocab, indexed in the same order
print("Vocabulary has %d distinct terms, examples below " % len(terms))
print(terms[500:550], '\n')

term_frequency_table = pd.DataFrame({'term': terms,'freq': list(np.array(doc_term_matrix.sum(axis=0)).reshape(-1))})
term_frequency_table = term_frequency_table.sort_values(by='freq', ascending=False).reset_index()

freq_df = pd.DataFrame(doc_term_matrix.todense(), columns = terms)
freq_df = freq_df.sum(axis=0)
freq_df = freq_df.sort_values(ascending=False)

# create the model
# Pick between NMF or LDA methods (don't know what they are, try whichever gives better results)
if NMF_or_LDA == 'nmf':
    model = NMF( init="nndsvd", n_components=num_topics ) 
else:
    model = LatentDirichletAllocation(n_components=num_topics, learning_method='online') 
    
# apply the model and extract the two factor matrices
W = model.fit_transform( doc_term_matrix ) #This matrix has docs as rows and k-topics as columns
H = model.components_ #This matrix has k-topics as rows and vocab as columns
print('Shape of W is', W.shape, 'docs as rows and', num_topics, 'topics as columns. First row below')
print(W[0].round(1))
print('\nShape of H is', H.shape, num_topics, 'topics as rows and vocab as columns. First row below')
print(H[0].round(1))

# Check which document belongs to which topic, and print value_count
topic_for_doc_df = pd.DataFrame(columns = ['article', 'topic', 'value'])
for i in range(W.shape[0]):
    a = W[i] 
    b = np.argsort(a)[::-1]
    temp_df = pd.DataFrame({'article': [i], 'topic':['Topic_'+str(b[0])], 'value': [a[b[0]]]})
    topic_for_doc_df = pd.concat([topic_for_doc_df, temp_df])

top_docs_for_topic_df = pd.DataFrame(columns = ['topic', 'doc_number', 'weight'])    
for i in range(W.shape[1]):
    topic = i
    temp_df = pd.DataFrame({'topic': ['Topic_'+str(i) for x in range(W.shape[0])], 
                            'doc_number':  list(range(W.shape[0])), 
                            'weight': list(W[:,i])})
    temp_df = temp_df.sort_values(by=['topic', 'weight'], ascending=[True, False])
    top_docs_for_topic_df = pd.concat([top_docs_for_topic_df, temp_df])
# Add text to the top_docs dataframe as a new column
top_docs_for_topic_df['text']=[raw_documents[i] for i in list(top_docs_for_topic_df.doc_number)] 
# Print top two docs for each topic
print('\nTop documents for each topic')
print(top_docs_for_topic_df.groupby('topic').head(2))

print('\n')
print('Topic number and counts of documents against each:')
print(topic_for_doc_df.topic.value_counts())

# Create dataframe with top-10 words for each topic
words_in_topics_df = pd.DataFrame(columns = ['topic', 'words', 'freq'])
for i in range(H.shape[0]):
    a = H[i] 
    b = np.argsort(a)[::-1]
    np.array(b[:top_n_words])
    words = [terms[i] for i in b[:top_n_words]]
    freq = [a[i] for i in b[:top_n_words]]
    temp_df = pd.DataFrame({'topic':'Topic_'+str(i), 'words': words, 'freq': freq})
    words_in_topics_df = pd.concat([words_in_topics_df, temp_df])

print('\n')
print('Top', top_n_words, 'words dataframe with weights')
print(words_in_topics_df.head(10))



# print as list
print('\nSame list as above as a list')
words_in_topics_list = words_in_topics_df.groupby('topic')['words'].apply(list)
lala =[]
for i in range(len(words_in_topics_list)):
    a = [list(words_in_topics_list.index)[i]]
    b = words_in_topics_list[i]
    lala = lala + [a+b]
    print(a + b) 
    
    
# Top terms
print('\nTop 10 most numerous terms:')
term_frequency_table.head(10)


freq_df

term_frequency_table.sort_values(by='freq', ascending=False)

final

pwd

newspaper.hot()
