==> Step 4 - NamedEntityRecognition.ipynb <==
# When on Google Colab, run this first, then Restart Kernel, and then load the libraries
!python -m spacy download en_core_web_lg

import joblib
import pandas as pd
import spacy
from tqdm import tqdm
nlp = spacy.load("en_core_web_lg")

# SAMPLE
# Download a spacy model for processing English
# !python -m spacy download en_core_web_lg
# nlp = spacy.load("en_core_web_sm")

# Process a sentence using the spacy model
doc = nlp('Windows 7 users should be aware by now that while a fair number of flaws addressed this month \
            by Microsoft affect Windows 7 systems, this operating system is no longer being supported with \
            security updates (unless you’re an enterprise taking advantage of Microsoft’s paid extended \
            security updates program, which is available to Windows 7 Professional and Windows 7 enterprise \
            users). Perhaps most troubling of these (CVE-2020-1301) is a remote code execution bug in SMB \
            capabilities built into Windows 7 and Windows Server 2008 systems — both operating systems that \
            Microsoft stopped supporting with security updates in January 2020. Microsoft Office and Excel get \
            several updates this month. Microsoft today released software patches to plug at least 129 security \
            holes in its Windows operating systems and supported software, by some accounts a record number of \
            fixes in one go for the software giant')
# doc = nlp(text)
# Display the entities found by the model, and the type of each.
print('{:<12}  {:}\n'.format('Entity', 'Type'))

# For each entity found...
for ent in doc.ents:
    
    # Print the entity text `ent.text` and its label `ent.label_`.
    print('{:<20}  {:}'.format(ent.text, ent.label_))

 
***
### Start here afresh with text sitting in a dataframe column

final = joblib.load('final_df3.pkl')
text = final.text.reset_index(drop=True)
text

df = pd.DataFrame(columns = ['entity_name', 'entity_type', 'doc_index'])

for n, t in enumerate(tqdm(text)):
    doc = nlp(t)
    for ent in doc.ents:
        ent_name = ent.text
        ent_label = ent.label_
        df = df.append({'entity_name': ent_name, 'entity_type': ent_label, 'doc_index': n}, ignore_index=True)

# Now the dataframe df contains all combinations of entity_name and entity_type and document.
# Count each entity and type only once for a document

df = df.drop_duplicates() 

# Now see how many documents does a given entity_name+entity_type combo appear in

entities = df.groupby(['entity_name', 'entity_type']).agg({'entity_name':'count'}).rename(columns={'entity_name':'COUNT'}).reset_index().sort_values(by = 'COUNT', ascending = False).reset_index(drop=True)

df

entities

entities.query('entity_type == "PERSON"')

print(list(entities.query('entity_type == "PERSON"').entity_name))

len(df)

df
